Superhirn PATCH by Florian Zeller: Explicit casts in MS1_feature_merger and FT_PeakDetectControler (size_t to int)
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@10080 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/TRANSFORMATIONS/FEATUREFINDER/SUPERHIRN/FT_PeakDetectController.C
@@ -133,7 +133,7 @@ void FT_PeakDetectController::start_scan_parsing_of_mzXML_file(Vec datavec){
   
   // create a new LC/MS:
   THIS_LC_MS = new LC_MS( name );
-  THIS_LC_MS->set_spectrum_ID( this->LC_MS_RUNS.size() );
+  THIS_LC_MS->set_spectrum_ID( (int) this->LC_MS_RUNS.size() );
   
   // start the mzXML reading process:
   FT_READER->read_mzXML_DATA(datavec);
